file structure and update packages
